Start of vfs, the Virtual File System.
Goal is to provide a transparent layer over filesystems so that all architectures can use the
same file io and directory structures.

For instance, a virtual file system could be created in Javascript that uses cookies for
storing path, file, and file contents information. Specialized writers, like a zip or tar
instance could be mounted on Root paths, and treated like any real filesystem.

Locking is being added to Input and Output to accomodate vfs.